finally

+++ src/main/java/com/fuzzyDucks/fms/Factories/impl/FileFactory.java
@@ -4,8 +4,10 @@ import com.fuzzyDucks.fms.Exceptions.NullDataException;
 import com.fuzzyDucks.fms.Factories.intf.IFileFactory;
 import com.fuzzyDucks.fms.File.File;
 import com.fuzzyDucks.fms.File.fileSchema.models.FileSchema;
+import com.fuzzyDucks.fms.File.impl.FileActions;
 import com.fuzzyDucks.fms.File.impl.FileServiceImpl;
 import com.fuzzyDucks.fms.File.intf.FileService;
+import com.sun.media.sound.InvalidDataException;
 
 import java.io.IOException;
 
@@ -15,30 +17,25 @@ public class FileFactory implements IFileFactory {
     public void doAction(String action, Object object) throws IOException, ClassNotFoundException {
         File file = (File) object;
         action = action.toLowerCase();
-        switch (action) {
-            case "import":
-                String path = file.getPath();
-                if (path == null || path.isEmpty()) {
-                    throw new NullDataException("Path is null or empty");
-                } else {
-                    new FileSchema(new java.io.File(file.getPath()));
-                }
-                break;
-            default:
-                throw new IllegalArgumentException("Invalid action");
+        if (FileActions.IMPORT.getValue().equals(action)) {
+            String path = file.getPath();
+            if (path == null || path.isEmpty()) {
+                throw new NullDataException("Path is null or empty");
+            } else {
+                new FileSchema(new java.io.File(file.getPath()));
+            }
+        } else {
+            throw new InvalidDataException("Invalid action");
         }
     }
     public void doAction(String action, String name,String type) throws IOException, ClassNotFoundException {
         action = action.toLowerCase();
-        switch (action) {
-            case "export":
-                fileService.exportFile(name,type);
-                break;
-            case "delete":
-                fileService.deleteFile(name, type);
-                break;
-            default:
-                throw new IllegalArgumentException("Invalid action");
+        if (FileActions.EXPORT.getValue().equals(action)) {
+            fileService.exportFile(name, type);
+        } else if (FileActions.DELETE.getValue().equals(action)) {
+            fileService.deleteFile(name, type);
+        } else {
+            throw new InvalidDataException("Invalid action");
         }
     }
